Improved file handling in man.py.
The main change was the use a context manager to open the CSV file.
With this change, we can remove the `finally` block and make sure that
the program will only try to close the file if the file was successfully
open.

The second change was the inclusion of code to handle
`FileNoteFoundError` and `PermissionErro` exceptions without crashing
the program.