Implemented `csv_to_contacts()` in `contact_utils.py`.
The function converts a contacts list from CSV into a list of `Contact`
objects.
This list can be serialized with pickle, json, or other format.

The `main.py` file was modified to use `csv_to_contacts()` to retrieve
the data form the CSV file as a list of `Contact` instances.