Implemented `contacts_to_pickle()` and `pickle_to_contacts()`.
The new methods will serialize and deserialize a list of Contact objets
to and from a pickle file.

The `main.py` was modified to use these methods and will create the file
`contacts.p` upon execution.
Since pickle files might be unsafe, I have not added it to the
repository.