Implemented `contacts_to_json()` and `json_to_contacts()`.
The methods server for serialization and deserialization of contacts
list using to and from JSON files.

The `main.py` was modified to utilize these functions to write and read
the JSON file.

Since JSON doesn't offer the risks of pickle, the generated JSON has
been included in the repo.

Also removed a few blank lines from the code, as they weren't really
making it easier to read the code.